Extract `gpath_bool` from `mod gstd::ops` path_bool-related nodes
