+++ editor/src/messages/tool/tool_messages/select_tool.rs
@@ -22,10 +22,10 @@ use crate::messages::tool::common_functionality::utility_functions::{resize_boun
 use bezier_rs::Subpath;
 use glam::DMat2;
 use graph_craft::document::NodeId;
+use graphene_std::path_bool::BooleanOperation;
 use graphene_std::renderer::Quad;
 use graphene_std::renderer::Rect;
 use graphene_std::transform::ReferencePoint;
-use graphene_std::vector::misc::BooleanOperation;
 use std::fmt;
 
 #[derive(Default)]
