+++ node-graph/gcore/src/vector/misc.rs
@@ -13,22 +13,6 @@ pub enum CentroidType {
 	Length,
 }
 
-#[derive(Default, Debug, Clone, Copy, PartialEq, Eq, serde::Serialize, serde::Deserialize, Hash, DynAny, specta::Type, node_macro::ChoiceType)]
-#[widget(Radio)]
-pub enum BooleanOperation {
-	#[default]
-	#[icon("BooleanUnion")]
-	Union,
-	#[icon("BooleanSubtractFront")]
-	SubtractFront,
-	#[icon("BooleanSubtractBack")]
-	SubtractBack,
-	#[icon("BooleanIntersect")]
-	Intersect,
-	#[icon("BooleanDifference")]
-	Difference,
-}
-
 pub trait AsU64 {
 	fn as_u64(&self) -> u64;
 }
