+++ node-graph/graph-craft/Cargo.toml
@@ -16,6 +16,7 @@ loading = ["serde_json"]
 # Local dependencies
 dyn-any = { workspace = true }
 graphene-core = { workspace = true }
+graphene-path-bool = { workspace = true }
 graphene-application-io = { workspace = true }
 
 # Workspace dependencies
