+++ node-graph/gstd/src/lib.rs
@@ -6,9 +6,10 @@ pub mod http;
 pub mod image_color_palette;
 pub mod raster;
 pub mod text;
-pub mod vector;
 #[cfg(feature = "wasm")]
 pub mod wasm_application_io;
 
 pub use graphene_application_io as application_io;
+pub use graphene_core::vector;
 pub use graphene_core::*;
+pub use graphene_path_bool as path_bool;
