+++ node-graph/interpreted-executor/Cargo.toml
@@ -14,6 +14,7 @@ graphene-std = { workspace = true }
 graph-craft = { workspace = true }
 wgpu-executor = { workspace = true }
 graphene-core = { workspace = true }
+graphene-path-bool = { workspace = true }
 dyn-any = { workspace = true }
 
 # Workspace dependencies
